juju-crashdump now handles broken or missing machines better.
We no longer fail if there are 0 machines, and we no longer raise an
Exception when a single machine has not finished deploying, or is
otherwise broken.

In both cases, we simply print a message and continue.

This addresses the case where we are running inside of an automated
testing tool, on an environment that may be broken. We want to gather as
much information as we can in that case.

This also addresses the issue where a user is trying to gather logs from
a broken environment.